web-api: refactored MessengerDirectMessage
