+++ lect-presenter-swing/src/main/java/org/lecturestudio/presenter/swing/view/SwingSlidesView.java
@@ -578,7 +578,7 @@ public class SwingSlidesView extends JPanel implements SlidesView {
 
 			if (message instanceof MessengerDirectMessage) {
 				MessengerDirectMessage directMessage = (MessengerDirectMessage) message;
-				String recipient = directMessage.getRecipient();
+				String recipient = directMessage.getRecipientId();
 
 				if (recipient.equals("organisers")) {
 					messageView.setPrivateText(dict.get("text.message.to.organisators"));
